Update conversion code.
